chore: move generated
